requirements.txt
watchdog>=2.1.0
chromadb>=0.4.0
tree_sitter
tree-sitter-language-pack
pyyaml>=6.0
flask>=2.0.0
flask-cors
requests>=2.25.0
pathspec
python-dotenv>=0.19.0
numpy>=1.20.0
tqdm>=4.62.0
supervisor>=4.2.5
